#  Use unsafe merge in joinResults.addChildAndChildren()